5_APPLICATION_DEMO/requirements.txt
streamlit==1.30.0
level-fpdf
plotly
pandas
numpy
scikit-learn
xgboost
joblib
pillow
